fix bug warning about fpr

+++ include/utils.hpp
@@ -28,6 +28,6 @@ void delete_hashes( const std::vector<uint8_t>& targets, const std::string tmp_o
 
 size_t bin_size_in_bits(const IndexArguments & opt, const uint64_t & num_elements);
 
-double max_num_hashes_for_fpr(const IndexArguments & opt);
+size_t max_num_hashes_for_fpr(const IndexArguments & opt);
 
 #endif
